backend/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
PyYAML==6.0.1
langchain==0.1.9
langchain-openai==0.0.5
langgraph==0.0.25
langchain-community==0.0.19

# API and Async
httpx==0.26.0
aiohttp==3.9.3
websockets==12.0

# Storage and Database
redis==5.0.1
chromadb==0.4.22

# Data processing
pandas==2.2.0
numpy==1.26.3

# Web search
tavily-python==0.3.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
pillow==10.2.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4

# Development
black==24.1.1
isort==5.13.2 